dh --with autotools_dev to refresh config.{sub,guess}
Accept patch from Ubuntu djview4 4.9-4ubntu3 to refresh
config.{sub,guess} using dh_autotools-dev.  This is needed by porters
to get config/config.{sub,guess} updated to bleeding-edge or custom
versions.  The patch also added a build dependency on autotools-dev.
That is not actually necessary here, as we are also using
dh-autoreconf, and autotools-dev is required by automake and libtool
which are in turn required by dh-autoreconf.  But lintian will
complain if the dependency is not listed.